memory: stm32-omi: Fix period_ps overflow
In both drivers hyperbus-stm32.c and stm32-ospi.c, period_ps is computed
using the following formula:

period_ps = NSEC_PER_SEC / (bus_freq / 1000)

In case bus freq < 15.259 Mhz, we got period_ps > 65535 which can't be
stored in a u16.
Convert period_ps from u16 to u32 to fix this issue.
Update stm32_omi_find_byp_cmd() and stm32_omi_dlyb_init() prototype
accordingly.

Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/557843
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
ACI: CIBUILD <MDG-smet-aci-builds@list.st.com>

+++ drivers/mtd/hyperbus/hyperbus-stm32.c
@@ -139,7 +139,7 @@ static int stm32_hyperbus_calibrate(struct hyperbus_device *hbdev)
 	struct stm32_omi *omi = hyperbus->omi;
 	void __iomem *regs_base = omi->regs_base;
 	u32 prescaler;
-	u16 period_ps = 0;
+	u32 period_ps = 0;
 	u8 window_len = 0;
 	int ret;
 	bool bypass_mode = false;
